start refactor
